updating for loop

+++ 0x1E-search_algorithms/0-linear.c
@@ -18,9 +18,9 @@ int linear_search(int *array, size_t size, int value)
 	if (array == NULL)
 		return (-1);
 
-	for (size_t i = 0; i < size; i++)
+	for (size_t i = 0; i < (int)size; i++)
 	{
-		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 
 		if (array[i] == value)
 			return (i);
